# docs/openapi/api-docs.yaml
openapi: 3.0.3
info:
  title: e-commerce-backend
  description: |
    API for managing a e-commerce.
    Supports customer management, products, categories.
  contact:
    email: carlosarroyoam@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.4.0-snapshot
servers:
  - url: http://localhost:3000/api/v1
    description: Localhost server url
...
paths:
...
  /users:
    get:
      tags:
        - user
      summary: Gets all users
      operationId: getUsers
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/SizeParam"
        - $ref: "#/components/parameters/SortParam"
        - name: firstName
          in: query
          required: false
          description: User's first name.
          schema:
            type: string
            maxLength: 64
        - name: lastName
          in: query
          required: false
          description: User's last name.
          schema:
            type: string
            maxLength: 64
        - name: email
          in: query
          required: false
          description: User's email address.
          schema:
            type: string
            maxLength: 64
        - name: isActive
          in: query
          required: false
          description: Whether the user is active or not.
          schema:
            type: boolean
        - name: startDate
          in: query
          required: false
          description: Start date for date range filter in the ISO format YYYY-MM-DD.
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: false
          description: End date for date range filter in the ISO format YYYY-MM-DD.
          schema:
            type: string
            format: date
        - name: userRoleId
          in: query
          required: false
          description: User role identifier.
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedUsersResponse"
        "422":
          $ref: "#/components/responses/UnprocesableEntity"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
      security:
        - BearerAuth: []
  /users/{user_id}:
    get:
      tags:
        - user
      summary: Gets a user by its id
      operationId: getUserById
      parameters:
        - name: user_id
          in: path
          description: ID of user to query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "422":
          $ref: "#/components/responses/UnprocesableEntity"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
      security:
        - BearerAuth: []

    delete:
      tags:
        - user
      summary: Deactivates a user by its id
      operationId: deactivateUser
      parameters:
        - name: user_id
          in: path
          description: ID of user to deactivate
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "422":
          $ref: "#/components/responses/UnprocesableEntity"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
      security:
        - BearerAuth: []
...
  /products:
    get:
      tags:
        - product
      summary: Gets all products
      operationId: getProducts
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/SizeParam"
        - $ref: "#/components/parameters/SortParam"
        - name: title
          in: query
          required: false
          description: Product's title.
          schema:
            type: string
            maxLength: 64
        - name: slug
          in: query
          required: false
          description: Product's slug.
          schema:
            type: string
            maxLength: 64
        - name: isFeatured
          in: query
          required: false
          description: Whether the product is featured or not.
          schema:
            type: boolean
        - name: isActive
          in: query
          required: false
          description: Whether the product is active or not.
          schema:
            type: boolean
        - name: startDate
          in: query
          required: false
          description: Start date for date range filter in the ISO format YYYY-MM-DD.
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: false
          description: End date for date range filter in the ISO format YYYY-MM-DD.
          schema:
            type: string
            format: date
        - name: categoryId
          in: query
          required: false
          description: Category identifier.
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedProductsResponse"
        "422":
          $ref: "#/components/responses/UnprocesableEntity"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /products/{product_id}:
    get:
      tags:
        - product
      summary: Gets a product by its id
      operationId: getProductById
      parameters:
        - name: product_id
          in: path
          description: ID of product to query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "422":
          $ref: "#/components/responses/UnprocesableEntity"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  schemas:
...
    PagedUsersResponse:
      type: object
      description: Paginated list of users response
      properties:
        items:
          type: array
          description: List of users
          items:
            $ref: "#/components/schemas/User"
        pagination:
          $ref: "#/components/schemas/PaginationDto"
...
    PagedProductsResponse:
      type: object
      description: Paginated list of products response
      properties:
        items:
          type: array
          description: List of products
          items:
            $ref: "#/components/schemas/Product"
        pagination:
          $ref: "#/components/schemas/PaginationDto"
...
    UserRole:
      type: object
      description: User role information
      properties:
        id:
          type: integer
          description: Unique identifier for the role
        type:
          type: string
          description: Type of user role
    User:
      type: object
      description: User information
      properties:
        id:
          type: integer
          description: Unique identifier for the user
        first_name:
          type: string
          description: User's first name
        last_name:
          type: string
          description: User's last name
        email:
          type: string
          description: User's email address
        is_active:
          type: boolean
          description: Indicates if user is active
        user_role:
          $ref: "#/components/schemas/UserRole"
        created_at:
          type: string
          format: date-time
          description: Timestamp when the user was created
        updated_at:
          type: string
          format: date-time
          description: Timestamp when the user was last updated
        deleted_at:
          type: string
          format: date-time
          description: Timestamp when the user was deleted (soft delete)
          nullable: true
...
    Category:
      type: object
      description: Category information
      properties:
        id:
          type: integer
          description: Unique identifier for the category
        title:
          type: string
          description: Category title
...
    Product:
      type: object
      description: Product information
      properties:
        id:
          type: integer
          description: Unique identifier for the product
        title:
          type: string
          description: Product title
        slug:
          type: string
          description: URL-friendly product identifier
        description:
          type: string
          description: Product description
        is_featured:
          type: boolean
          description: Indicates if product is featured
        is_active:
          type: boolean
          description: Indicates if product is active
        category:
          $ref: "#/components/schemas/Category"
        properties:
          type: array
          items:
          $ref: "#/components/schemas/ProductPropertyValue"
        variants:
          type: array
          items:
            $ref: '#/components/schemas/ProductVariant'
        created_at:
          type: string
          format: date-time
          description: Timestamp when the product was created
        updated_at:
          type: string
          format: date-time
          description: Timestamp when the product was last updated
        deleted_at:
          type: string
          format: date-time
          description: Timestamp when the product was deleted (soft delete)
          nullable: true
    Property:
      type: object
      description: Property information for products
      properties:
        id:
          type: integer
          description: Unique identifier for the property
        title:
          type: string
          description: Property title (e.g., Brand, Model)
    ProductPropertyValue:
      type: object
      description: Product property value information
      properties:
        id:
          type: integer
          description: Unique identifier for the Product property value
        value:
          type: string
          description: Product property value
        property:
          $ref: "#/components/schemas/Property"
    ProductVariant:
      type: object
      description: Product variant information
      properties:
        id:
          type: integer
          description: Unique identifier for the variant
        sku:
          type: string
          description: Stock keeping unit
        price:
          type: number
          description: Variant price
        compared_at_price:
          type: number
          description: Original price for comparison
        cost_per_item:
          type: number
          description: Cost per item
        quantity_on_stock:
          type: integer
          description: Available stock quantity
        attributes:
          $ref: "#/components/schemas/VariantAttributeValue"
        images:
          $ref: "#/components/schemas/VariantImage"
    Attribute:
      type: object
      description: Attribute information
      properties:
        id:
          type: integer
          description: Unique identifier for the attribute
        title:
          type: string
          description: Attribute title (e.g., Color, Storage)
    VariantAttributeValue:
      type: object
      description: Variant attribute value information
      properties:
        id:
          type: integer
          description: Unique identifier for the Variant attribute value
        value:
          type: string
          description: Variant attribute value
        attribute:
          $ref: "#/components/schemas/Attribute"
    VariantImage:
      type: object
      description: Variant image information
      properties:
        id:
          type: integer
          description: Unique identifier for the image
        url:
          type: string
          description: Image URL
        variant_id:
          type: integer
          description: Foreign key to variants table
...
    PaginationDto:
      type: object
      description: Pagination information for list endpoints
      properties:
        page:
          type: integer
          description: Current page number
        size:
          type: integer
          description: Number of items per page
        total_items:
          type: integer
          description: Total number of items
        total_pages:
          type: integer
          description: Total number of pages
    AppExceptionDto:
      type: object
      description: Standard error response returned when an application exception occurs.
      properties:
        message:
          type: string
          description: Human-readable error message.
        error:
          type: string
          description: Short error identifier or error type.
        status:
          type: integer
          description: HTTP status code associated with the error.
        details:
          type: object
          description: Additional validation or error details.
          additionalProperties:
            type: string
        path:
          type: string
          description: API endpoint path where the error occurred.
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the error was generated.
  parameters:
    SizeParam:
      name: size
      in: query
      description: Pagination size items
      schema:
        type: integer
        default: 50
    PageParam:
      name: page
      in: query
      description: Pagination page number
      schema:
        type: integer
        default: 1
    SortParam:
      name: sort
      in: query
      required: false
      description: Sorting criteria in the format field,asc or field,desc.
      schema:
        type: string
        default: id,asc
...
  responses:
...
    NoContent:
      description: No content
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AppExceptionDto"
...
    UnprocesableEntity:
      description: Unprocessable entity
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AppExceptionDto"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AppExceptionDto"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AppExceptionDto"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AppExceptionDto"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT